+++ lib/ex_aliyun_ots/application.ex
@@ -22,33 +22,53 @@ defmodule ExAliyunOts.Application do
     Supervisor.start_link(children, strategy: :one_for_one)
   end
 
+  def http_name() do
+    __MODULE__.Finch
+  end
+
   defp child_spec(_app, _enable_tunnel, []) do
     raise ExAliyunOts.RuntimeError, "Not found any instances configuration"
   end
 
   defp child_spec(app, true, instance_keys) do
+    {instances, pools} = load_instances(app, instance_keys)
+
     [
-      worker(ExAliyunOts.Config, [load_instances(app, instance_keys)]),
+      child_spec_http_client(pools),
+      worker(ExAliyunOts.Config, [instances]),
       worker(ExAliyunOts.Tunnel.Registry, []),
       supervisor(ExAliyunOts.Tunnel.DynamicSupervisor, [])
     ]
   end
 
   defp child_spec(app, false, instance_keys) do
+    {instances, pools} = load_instances(app, instance_keys)
+
     [
-      worker(ExAliyunOts.Config, [load_instances(app, instance_keys)])
+      child_spec_http_client(pools),
+      worker(ExAliyunOts.Config, [instances])
     ]
   end
 
+  defp child_spec_http_client(pools) do
+    {
+      Finch,
+      name: http_name(), pools: pools
+    }
+  end
+
   defp load_instances(app, instance_keys) do
-    Enum.reduce(instance_keys, [], fn instance_key, acc ->
-      [
-        {
-          instance_key,
-          Map.merge(%Instance{}, Enum.into(Application.get_env(app, instance_key), %{}))
-        }
-        | acc
-      ]
+    Enum.reduce(instance_keys, {[], %{}}, fn instance_key, {instances, pools_map} ->
+      instance = struct(Instance, Application.get_env(app, instance_key))
+      instances = [{instance_key, instance} | instances]
+
+      pools_map =
+        Map.put(pools_map, instance.endpoint,
+          size: instance.pool_size || 100,
+          count: instance.pool_count || 1
+        )
+
+      {instances, pools_map}
     end)
   end
 end