+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v0.9.0 (2020-11-17)
+
+* [Http] Use Tasla with Finch adapter
+
+## v0.8.0 (2020-11-13)
+* [Table] Add `ExAliyunOts.compute_split_points_by_size/2`
+* [Table] Fix `shard_splits` of DescribeTable response in proper decoded format
+* [SearchIndex] Add ParallelScan functions
+	- Implement standard api `ExAliyunOts.compute_splits/3`
+	- Implement standard api `ExAliyunOts.parallel_scan/4`
+	- Implement built-in functions `ExAliyunOts.iterate_parallel_scan/5`, `ExAliyunOts.iterate_parallel_scan/7` and
+    `ExAliyunOts.stream_parallel_scan/4` to make general use cases simple
+
 ## v0.7.1 (2020-09-15)
 * Enhance condition expression and simplify some code about it